a/requirements.txt
fastapi
uvicorn
python-multipart
pandas
openpyxl
pdfplumber
supabase
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
jinja2
aiofiles
gunicorn